test: prototype with assistant
