+++ src/pages/ProfilesExplorerView/components/SceneExploreServiceFlameGraph/components/SceneFunctionDetailsPanel/components/CodeContainer/ui/Code.tsx
@@ -1,7 +1,7 @@
 import { css, cx } from '@emotion/css';
-import { GrafanaTheme2, rangeUtil } from '@grafana/data';
-import { Button, LinkButton, Modal, Spinner, useStyles2 } from '@grafana/ui';
-import React, { useState } from 'react';
+import { GrafanaTheme2 } from '@grafana/data';
+import { LinkButton, Spinner, useStyles2 } from '@grafana/ui';
+import React from 'react';
 
 import { AIButton } from '../../../../../../../components/SceneAiPanel/components/AiButton/AIButton';
 import { buildUnitFormatter } from '../../../domain/buildUnitFormatter';
@@ -54,27 +54,8 @@ export const Code = ({
     [0, 0]
   );
 
-  const [prompt, setPrompt] = useState('');
-
-  const clearPrompt = () => setPrompt('');
-
   return (
     <div data-testid="function-details-code-container">
-      {
-        <Modal title="Prompt Example" isOpen={prompt !== ''} onDismiss={clearPrompt} onClickBackdrop={clearPrompt}>
-          <pre>{prompt}</pre>
-          <Modal.ButtonRow>
-            <Button
-              onClick={() => {
-                codeRelate(prompt);
-              }}
-            >
-              Investigate
-            </Button>
-            <Button onClick={clearPrompt}>Clear prompt</Button>
-          </Modal.ButtonRow>
-        </Modal>
-      }
       <div className={styles.container}>
         <div className={styles.header}>
           <div className={styles.breakdownLabel}>
@@ -91,27 +72,7 @@ export const Code = ({
                 fill="text"
                 icon={'search-plus'}
                 onClick={() => {
-                  const params = new URLSearchParams(document.URL);
-                  const serviceName = params.get('var-serviceName');
-                  const profileMetricId = params.get('var-profileMetricId');
-                  const shortProfileId = profileMetricId?.split(':').slice(0, 2).join('/');
-
-                  const from = params.get('from');
-                  const to = params.get('to');
-
-                  const timeRange = !from || !to ? 'last 30 minutes' : rangeUtil.describeTimeRange({ from, to });
-
-                  const prompt = [
-                    `Through a Profiles Drilldown query of the \`${shortProfileId}\` profile `,
-                    `for \`{service_name="${serviceName}"\` in the ${timeRange} time range, `,
-                    `I found the following source code:\n\n`,
-                    '```\n',
-                    functionCode,
-                    '\n```\n\n',
-                    'Can you find relevant observability data from those lines of code (traces, metrics, or logs)?',
-                  ].join('');
-
-                  setPrompt(prompt);
+                  codeRelate(functionCode);
                 }}
               >
                 Co(de) Relate
